Move default tileset parsing to Mods.Common.
