+++ OpenRA.Mods.Common/Traits/World/EditorCursorLayer.cs
@@ -12,6 +12,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using OpenRA.Graphics;
+using OpenRA.Mods.Common.Terrain;
 using OpenRA.Traits;
 
 namespace OpenRA.Mods.Common.Traits
