+++ OpenRA.Mods.Common/Widgets/Logic/Editor/TileSelectorLogic.cs
@@ -13,8 +13,8 @@ using System;
 using System.IO;
 using System.Linq;
 using OpenRA.Graphics;
+using OpenRA.Mods.Common.Terrain;
 using OpenRA.Mods.Common.Traits;
-using OpenRA.Traits;
 using OpenRA.Widgets;
 
 namespace OpenRA.Mods.Common.Widgets.Logic
