+++ OpenRA.Mods.Common/Widgets/TerrainTemplatePreviewWidget.cs
@@ -12,9 +12,9 @@
 using System;
 using System.IO;
 using OpenRA.Graphics;
+using OpenRA.Mods.Common.Terrain;
 using OpenRA.Mods.Common.Traits;
 using OpenRA.Primitives;
-using OpenRA.Traits;
 using OpenRA.Widgets;
 
 namespace OpenRA.Mods.Common.Widgets
